refactored the jpa module structure
