+++ src/main/java/vaultiq/session/jpa/session/repository/VaultiqSessionRepository.java
@@ -1,11 +1,10 @@
-package vaultiq.session.jpa.repository;
+package vaultiq.session.jpa.session.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import vaultiq.session.jpa.config.VaultiqJpaAutoConfiguration;
-import vaultiq.session.jpa.model.JpaVaultiqSession;
+import vaultiq.session.jpa.session.model.JpaVaultiqSession;
 
 import java.util.List;
-import java.util.Optional;
 
 /**
  * Instantiated only when the property vaultiq.session.persistence.jpa.enabled is set to true.
